Maze generation pt. 2
The generation function now backtracks, so if it is stuck in an area, it back tracks and continues on the path. This allows for a real maze, with there being a correct path between a defined start and end, and the incorrect branches that were backtracked being incorrect choices in the maze.